requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
bidict==0.22.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.3
Flask-CLI==0.4.0
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
flask-restx==0.5.1
flask-sock==0.5.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
livereload==2.6.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyCLI==2.0.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-dotenv==0.19.1
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
simple-websocket==0.5.1
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.1.0
